Describe SCSS workflow
Updated a bit to be more consistent with https://github.com/TelescopeJS/Telescope/wiki/CSS. All of the wiki page should probably just be moved over into the docs, but this seemed like a fine first start so that it's at least not out of date. I don't have a local instance of the homepage running, but this seems like it should all be fine.